+++ .github/workflows/pr-main_l2.yaml
@@ -31,6 +31,13 @@ jobs:
         with:
           components: rustfmt, clippy
 
+      # Required for build scripts
+      - name: Install solc
+        uses: pontem-network/get-solc@master
+        with:
+          version: v0.8.29
+          token: ${{ secrets.GITHUB_TOKEN || '' }}
+
       - name: Run cargo check
         run: cargo check --workspace
