+++ crates/l2/contracts/bin/deployer/main.rs
@@ -149,8 +149,7 @@ async fn deploy_contracts(
     let on_chain_proposer_deployment = deploy_with_proxy(
         opts.private_key,
         eth_client,
-        &opts.contracts_path.join("solc_out"),
-        "OnChainProposer.bin",
+        &opts.contracts_path.join("solc_out/OnChainProposer.bin"),
         &salt,
     )
     .await?;
@@ -168,8 +167,7 @@ async fn deploy_contracts(
     let bridge_deployment = deploy_with_proxy(
         opts.private_key,
         eth_client,
-        &opts.contracts_path.join("solc_out"),
-        "CommonBridge.bin",
+        &opts.contracts_path.join("solc_out/CommonBridge.bin"),
         &salt,
     )
     .await?;
@@ -188,8 +186,7 @@ async fn deploy_contracts(
         let sequencer_registry_deployment = deploy_with_proxy(
             opts.private_key,
             eth_client,
-            &opts.contracts_path.join("solc_out"),
-            "SequencerRegistry.bin",
+            &opts.contracts_path.join("solc_out/SequencerRegistry.bin"),
             &salt,
         )
         .await?;
