+++ crates/l2/sdk/contract_utils/src/lib.rs
@@ -1,0 +1,5 @@
+mod compile;
+mod deps;
+
+pub use compile::*;
+pub use deps::*;
