+++ crates/l2/tee/quote-gen/service.nix
@@ -21,6 +21,43 @@ let
     sha256 = "sha256-HG2cCnktfHsKV0s4XW83gU3F57gaTljL9KNSuG6bnQs";
   };
   inherit (import gitignoreSrc { inherit (pkgs) lib; }) gitignoreSource;
+
+  solc_0_8_29 = pkgs.stdenv.mkDerivation {
+    name = "solc-0.8.29";
+    src = pkgs.fetchurl {
+      url = "https://github.com/ethereum/solidity/releases/download/v0.8.29/solc-static-linux";
+      sha256 = "sha256-GNQYpA3ATRdlaxtcins1z7q4lCtR840AXVtZ6KpmN+A=";
+    };
+    phases = [ "installPhase" ];
+    installPhase = ''
+      mkdir -p $out/bin
+      cp $src $out/bin/solc
+      chmod +x $out/bin/solc
+    '';
+  };
+
+  openzeppelin = pkgs.fetchgit {
+    url = "https://github.com/OpenZeppelin/openzeppelin-contracts-upgradeable.git";
+    rev = "v5.3.0";
+    sha256 = "sha256-DliCWCetMsVtU7oNBsbk9shiC6MymX1X+qejkNNWpRE";
+    fetchSubmodules = true;
+  };
+
+  sp1_contracts = pkgs.fetchgit {
+    url = "https://github.com/succinctlabs/sp1-contracts.git";
+    rev = "v4.0.0";
+    sha256 = "sha256-QqgIISFoN1bnc2Bz0T6MXiP6hOg7J5+tEDFmOgRh6qM=";
+    fetchSubmodules = true;
+  };
+
+  contracts = pkgs.runCommand "contracts" {} ''
+    mkdir -p $out/lib/openzeppelin-contracts-upgradeable
+    cp -r ${openzeppelin}/* $out/lib/openzeppelin-contracts-upgradeable
+
+    mkdir -p $out/lib/sp1-contracts
+    cp -r ${sp1_contracts}/* $out/lib/sp1-contracts
+  '';
+
 in
 let
   quoteGen = rustPlatform.buildRustPackage rec {
@@ -44,8 +81,13 @@ let
     nativeBuildInputs = [
       pkgs.pkg-config
       rustPlatform.cargoSetupHook
+      solc_0_8_29
     ];
-    env.OPENSSL_NO_VENDOR = 1;
+
+    env = {
+      OPENSSL_NO_VENDOR = 1;
+      CONTRACTS_PATH = "${contracts}/lib";
+    };
   };
 in
 {
